Add files via upload
The index.html is the html file for the web page asked to design and the index.css is it's stylesheet. The input.html is the html file for displaying input on the screen. The input.js is the Javascript file which prompts the user for the input and displays the input on the screen.